build(hibernate): agrega dependencias de Hibernate
- Implementa `MovieRepository`.
- Agrega logs.

+++ src/main/java/io/redbee/academy/persistence/hibernate/repository/MovieRepository.java
@@ -9,7 +9,7 @@ public interface MovieRepository {
 
     List<Movie> getAll();
 
-    Optional<Movie> find(Long id);
+    Optional<Movie> find(Integer id);
 
     List<Movie> find(String title);
